bugfix(DPAV-2220): fix loading spinner getting stuck after focus area changes
- Fix race condition where map idle event was missed if map was already idle
- Extract MapLoadingOverlay into separate component
- Simplify DrawingContext to use requestAnimationFrame instead of idle event
- Add 3s timeout fallback to prevent spinner from getting stuck indefinitely
