+++ frontend/audit-ci.jsonc
@@ -1,5 +1,5 @@
 {
     "$schema": "https://github.com/IBM/audit-ci/raw/main/docs/schema.json",
     "low": true,
-    "allowlist": []
+    "allowlist": [],
 }
